Add flag to calico windows uninstall script and pass it from hostprocess install
The calico windows uninstall script would not run fully when it detects
it is running in a hostprocess container, but it is called by the
hostprocess installation to clean up a previous manual install, so
it needs to run fully in this special case. Fixing it by adding a flag
to indicate this situation.

+++ node/windows-packaging/CalicoWindows/uninstall-calico.ps1
@@ -13,7 +13,8 @@
 # limitations under the License.
 param
 (
-    [bool][parameter(Mandatory=$false)]$ExceptUpgradeService = $false
+    [bool][parameter(Mandatory=$false)]$ExceptUpgradeService = $false,
+    [bool][parameter(Mandatory=$false)]$CleanupForHostProcessInstall = $false
 )
 
 ipmo "$PSScriptRoot\libs\calico\calico.psm1" -Force
@@ -25,8 +26,10 @@ Test-CalicoConfiguration
 $ErrorActionPreference = 'SilentlyContinue'
 
 # If running in a hostprocess container, remove Calico CNI if installed.
-# Skip the rest of the logic that applies to manual installations only.
-if (($env:CONTAINER_SANDBOX_MOUNT_POINT) -and ($env:CALICO_NETWORKING_BACKEND -NE "none"))
+# Skip the rest of the logic that applies to manual installations only,
+# unless this is being run by the hostprocess installation itself (which
+# will pass the -CleanupForHostProcessInstall param).
+if (($env:CONTAINER_SANDBOX_MOUNT_POINT) -and ($env:CALICO_NETWORKING_BACKEND -NE "none") -and (-Not $CleanupForHostProcessInstall))
 {
     if ($env:CALICO_NETWORKING_BACKEND -NE "none") {
         Remove-CNIPlugin
